+++ tests/nv/test_ex_routes.py
@@ -263,6 +263,7 @@ class TestRoutes(unittest.TestCase):
         self.assertRoute('_ex_route_file', ['file', 'f'])
         self.assertRoute('_ex_route_global', ['global', 'g'])
         self.assertRoute('_ex_route_help', ['help', 'h'])
+        self.assertRoute('_ex_route_history', ['history', 'his'])
         self.assertRoute('_ex_route_let', ['let '])
         self.assertRoute('_ex_route_move', ['move', 'm'])
         self.assertRoute('_ex_route_new', ['new'])